refactor: コントローラーの中身を全てモデルに移行.
